Bug 1530517 - build the UI as part of docker image building
