+++ infrastructure/builder/src/generate/generators/entry-script.js
@@ -29,10 +29,10 @@ exports.tasks.push({
     //  - starts nginx to serve that rootUrl-relative output
     procs['references/web'] = 'exec sh infrastructure/references/references.sh';
 
-    // the ui/web process runs `yarn build` before starting nginx to serve
-    // the resulting, built content
-    procs['ui/web'] = 'cd /app/ui && GENERATE_ENV_JS=1 yarn build && ' +
-      ' nginx -c /app/ui/web-ui-nginx-site.conf -g \'daemon off;\'';
+    // the ui/web process sets up static/env.js to relay configuration values to
+    // the frontend, then runs nginx.
+    procs['ui/web'] = 'node /app/ui/generate-env-js.js /app/ui/build/static/env.js && ' +
+      'nginx -c /app/ui/web-ui-nginx-site.conf -g \'daemon off;\'';
 
     const entrypointScript = []
       .concat([
