+++ ui/.neutrinorc.js
@@ -2,19 +2,10 @@ const merge = require('deepmerge');
 const copy = require('@neutrinojs/copy');
 const { join, resolve } = require('path');
 const fs = require('fs');
+const generateEnvJs = require('./generate-env-js');
 
 const DEFAULT_PORT = 5080;
 const STATIC_DIR = join(__dirname, 'src/static');
-const ENV_DEFAULTS = {
-  APPLICATION_NAME: 'Taskcluster',
-  TASKCLUSTER_ROOT_URL: 'https://tc.example.com',
-  GRAPHQL_ENDPOINT: '/graphql',
-  GRAPHQL_SUBSCRIPTION_ENDPOINT: '/subscription',
-  DOCS_ONLY: false,
-  UI_LOGIN_STRATEGY_NAMES: '',
-  GA_TRACKING_ID: '',
-  SENTRY_DSN: '',
-};
 const port = process.env.PORT || DEFAULT_PORT;
 
 require('@babel/register')({
@@ -175,29 +166,16 @@ module.exports = {
         ],
       });
 
-      // Generate env.js if it doesn't exist
+      // Generate env.js, combining env vars into the build, when
+      // GENERATE_ENV_JS is set
+      const envJs = join(STATIC_DIR, 'env.js');
       if (process.env.GENERATE_ENV_JS) {
-        const envJs = `window.env = ${
-          JSON.stringify(
-            Object.keys(ENV_DEFAULTS).reduce((acc, curr) => {
-              acc[curr] = process.env[curr] || ENV_DEFAULTS[curr];
-
-              return acc;
-            }, {}), null, 2)
-        };`;
-
-        if (!fs.existsSync(STATIC_DIR)){
-          fs.mkdirSync(STATIC_DIR);
-        }
-
-        if (!fs.existsSync(join(STATIC_DIR, 'env.js'))){
-          fs.writeFileSync(join(STATIC_DIR, 'env.js'), envJs, 'utf8');
-        }
+        generateEnvJs(envJs);
       } else {
         // just so that we never end up accidentally including something
         // in a production build
-        if (fs.existsSync(join(STATIC_DIR, 'env.js'))) {
-          fs.unlinkSync(join(STATIC_DIR, 'env.js'));
+        if (fs.existsSync(envJs)) {
+          fs.unlinkSync(envJs);
         }
       }
     },
